+++ MD/app/build.gradle.kts
@@ -56,9 +56,11 @@ dependencies {
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    implementation ("androidx.activity:activity-ktx:1.9.3")
+    implementation (libs.androidx.activity.ktx)
     implementation (libs.circleimageview)
     implementation(libs.glide)
+    implementation ("com.github.yalantis:ucrop:2.2.8")
+
     implementation("com.squareup.retrofit2:retrofit:2.9.0")
     implementation("com.squareup.retrofit2:converter-gson:2.9.0")
     implementation("com.squareup.okhttp3:logging-interceptor:4.11.0")
