+++ MD/app/src/main/AndroidManifest.xml
@@ -49,6 +49,12 @@
             android:name=".MainActivity"
             android:exported="false"
             android:label="@string/app_name" />
+
+        <activity
+            android:name="com.yalantis.ucrop.UCropActivity"
+            android:screenOrientation="fullSensor"
+            android:theme="@style/Theme.AppCompat.Light.NoActionBar"
+            tools:ignore="DiscouragedApi" />
     </application>
 
 </manifest>
